[ports-mingw] zxing-cpp: update to 2.2.1
This is a fix for a c++ ABI breakage of the 2.2.0 release. Only of relevance if people want to run an executable that compiled with dynamically linking against version 2.1.0 and is still supposed to run with a new dll (of version 2.2.1).

Source: msys2/MINGW-packages@a93990c3bdb6b4c208ef8bc9eaad7c58be3c2553