[ports-mingw] Add a msys2 field for ignoring CVEs
Not sure if we'll use it really, but let's give it a try.

This just ignores a random CVE that is missing version range info
and thus is a false positive.

We might exand it in the future to make it a mapping and include a
reason, so we can show it on the website.

Source: msys2/MINGW-packages@9ab3e3a4f7781f62ba32455e3e6fe17b6b88ab97