[ports] CI: move dll checks into per-package loop
This way the package, and its dependencies, are installed, so there
shouldn't be any missing DLL dependencies.

This is closer to what MINGW-packages does (with ntldd), but the scripts
could most likely stand to be synced to be closer.

Source: msys2/MSYS2-packages@8217d1000cbc6b97a7a603fa292ae5395441ad85