[ports-mingw] python-installer: Zero lpReserved2 and cbReserved2 fields in simple_launcher STARTUPINFOW
simple_launcher currently copies the values of lpReserved2 and cbReserved2 from
the parent process to the child. In some cases, these fields are used to pass
information. For example, the MSVC C runtime uses the fields as part of its
_spawn implementation to pass file descriptor information to the child. However,
this information is invalid in the grand-child proces.

This issue is a cause of a bug when CMake attempts to launch a python-installer
child process:

  * https://gitlab.kitware.com/cmake/cmake/-/issues/25996

This patch corrects the issue by applying a patch to zero the fields.

Source: msys2/MINGW-packages@a9b746da714e441244a14dde72368ac7c3e4bd7b