[ports-mingw] octave: work around issue with MinGW makeinfo
When trying to use the MinGW version of `makeinfo` displaying a help
message fails with an error like the following:
```
Can't call method "name" on an undefined value at makeinfo line 508.
```
That is a regression.

Restore the work-around that was used previously and was removed in
https://github.com/msys2/MINGW-packages/commit/cacaa18d5ddc6a8197974f2e795a96e1eb3224cb

Source: msys2/MINGW-packages@210e6ac7afe65c7c56ad686c558a81d99abfdc1e