[ports] perl: remove renames in comments
This is since this used to be a separate msys.sh file
which was synced back.

Let's keep things simple and only patch what we need to.

TODO: check if the other changes in that file are really needed,
or if they just diverged since the split for no good reason.

Source: msys2/MSYS2-packages@74504c8e73e0248b13aa882502cd0d6fb1af139d