[ports] filesystem: Remove /etc/bash.bash_logout (#5221)
* filesystem: Remove /etc/bash.bash_logout

By adding `Hold=always` in ~/.minttyrc, it is possible to keep MinTTY open
after all processes in it have terminated, which is useful for examination
of outputs. However this script used to clear the screen, leaving a blank
window, and causing confusion.

Another reason for removing this script is that it's not found on fresh
installation of Debian or Ubuntu.

Reference: https://superuser.com/questions/1262655/mintty-how-to-keep-my-output-visible-after-processes-terminate/1263001#1263001
Signed-off-by: LIU Hao <lh_mouse@126.com>

* Update PKGBUILD

---------

Signed-off-by: LIU Hao <lh_mouse@126.com>
Co-authored-by: Christoph Reiter <reiter.christoph@gmail.com>

Source: msys2/MSYS2-packages@71d24745afd0790e36bb7b732a7ad2c0e74caa9e