[ports] genromfs: downgrade to 0.5.2 - the most recent true release (#5234)
* genromfs: downgrade to 0.5.2 - the most recent true release

The current package version is 0.5.7. This version number comes from a tag in
the GitHub repository at commit 2bdc5f9 - a commit made in 2009. However, this
tag was never released as a true package release and the commit has the label:

  Update to version 0.5.7 (not released)

Repology reports that the vast majority of package distributions are shipping
the 0.5.2 release, which is the most recent true release made in 2007:

  https://repology.org/project/genromfs/versions

One bug that was found with the 0.5.7 non-release is that the genromfs tool
opens file handles without then closing them which eventually causes a "Too
many open files" error. A patch has been submitted to fix the issue here:

  https://github.com/chexum/genromfs/pull/5

However, it seems unlikely that acceptance of this patch, or a new release are
going to be forthcoming because no commits have been added to the GitHub
repository since 2011.

Therefore, this patch sets the version number to 0.5.2.

Signed-off-by: Joel Holdsworth <jholdsworth@nvidia.com>

* bump epoch

---------

Signed-off-by: Joel Holdsworth <jholdsworth@nvidia.com>
Co-authored-by: Christoph Reiter <reiter.christoph@gmail.com>

Source: msys2/MSYS2-packages@6a9da6ebce00960c8674d0de797b18a67a718f8f