[ports] serf: rebuild and adjust shared lib name
With newer scons we otherwise get "liblib" prefixes.
It's not clear if that is a msys specific scons issue, or if
this has changed in scons.

It looks like trunk serf has updated all this code, so maybe it
is fixed there, we'll see.

Source: msys2/MSYS2-packages@57f40bfb5da9f75c6f6d57fa87bec339c57101ea