[ports-mingw] Revert "winpthreads-stub: Add package to be link without pthreads library (#24413)"
This reverts commit 7217dc80453f85bd51d33c51cdc041fd4aba2e11.

Turns out this gets pulled in now instead of winpthreads, because our main package is called winpthreads-git and the in case of multiple provides one of them wins.

So we'll have to revert this, and remove the "-git" prefix from the main package if we want this to work.

Source: msys2/MINGW-packages@727b2084226f1425229d2cdcbb18f17189cef3f8