[ports-mingw] CI: remove procdump setup
It was added in 6ff1a1f8a631ab9b243a8, but we never really needed
it since then.

Since MS is currently down and this causes CI to fail, it's a good time
to remove it.

If we need it again it's always in the git history..

Source: msys2/MINGW-packages@a1e12d602dc80ee0df3f3f0658dc8a9e9e8a9532