[ports-mingw] brotli: split Python package into separate PKGBUILD
This seems to be required, because otherwise the normal brotli
package would makedepend on python-installer and the check step
in the CI would complain about the strip option not being
disabled. However, you don't want to do that for the normal
brotli package.

Source: msys2/MINGW-packages@9b344635687ddaf0703725e32965f212ce40208e