[ports-mingw] msmpi: Fix value of MPI_DATATYPE_NULL
The value was incorrectly converted in the header and module for the changes in 66462c2fcd927e27acbd561a56bb5638fc6d2f6a.

Use the correct file extension for a Fortran source file that needs to be preprocessed (and remove a flag that is no longer needed in the compile command).

Add a check function to make it easier to run the tests. This requires that Microsoft MPI is installed on the system where the tests are run.

Source: msys2/MINGW-packages@d62144100f378f5f0926822d91f28432e8088c1b